feat: recycler view adicionado e parcialmente configurado na home fragment. reorganização de diretórios
