+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.PetsUp"
         tools:targetApi="31">
         <activity
-            android:name=".activity.MainActivity"
+            android:name=".ui.activity.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -21,10 +21,10 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".activity.SignupActivity" />
-        <activity android:name=".activity.LoginActivity" />
-        <activity android:name=".activity.BottomMenuActivity" />
-        <activity android:name=".activity.PetshopDetailActivity" />
+        <activity android:name=".ui.activity.SignupActivity" />
+        <activity android:name=".ui.activity.LoginActivity" />
+        <activity android:name=".ui.activity.BottomMenuActivity" />
+        <activity android:name=".ui.activity.PetshopDetailActivity" />
 
         <meta-data
             android:name="preloaded_fonts"
