+++ app/src/main/java/com/petsup/ui/activity/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.petsup.activity
+package com.petsup.ui.activity
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
