+++ app/src/main/java/com/petsup/ui/activity/PetGenderActivity.kt
@@ -1,4 +1,4 @@
-package com.petsup.activity
+package com.petsup.ui.activity
 
 import android.content.Intent
 import android.os.Bundle
@@ -7,23 +7,18 @@ import android.widget.Button
 import android.widget.ImageView
 import androidx.appcompat.app.AppCompatActivity
 import com.petsup.R
-import com.petsup.`object`.ValidationObject
-import com.petsup.fragment.PetListFragment
 
-class PetNameActivity : AppCompatActivity() {
+class PetGenderActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_pet_name)
+        setContentView(R.layout.activity_pet_gender)
     }
 
-    fun confirm(view: View) {
-        val confirmButton: Button = findViewById(R.id.confirm_button)
-        confirmButton.setOnClickListener {
-            val isValid = ValidationObject.petNameValidation(findViewById(R.id.name_edit_text))
-            if (isValid) {
-                val intent = Intent(this, PetListFragment::class.java)
-                startActivity(intent)
-            }
+    fun next(view: View) {
+        val continueButton: Button = findViewById(R.id.continue_button)
+        continueButton.setOnClickListener {
+            val intent = Intent(this, PetNameActivity::class.java)
+            startActivity(intent)
         }
     }
