+++ app/src/main/java/com/petsup/ui/activity/PetNameActivity.kt
@@ -1,4 +1,4 @@
-package com.petsup.activity
+package com.petsup.ui.activity
 
 import android.content.Intent
 import android.os.Bundle
@@ -7,18 +7,23 @@ import android.widget.Button
 import android.widget.ImageView
 import androidx.appcompat.app.AppCompatActivity
 import com.petsup.R
+import com.petsup.ui.`object`.ValidationObject
+import com.petsup.ui.fragment.PetListFragment
 
-class PetSpeciesActivity : AppCompatActivity() {
+class PetNameActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_pet_species)
+        setContentView(R.layout.activity_pet_name)
     }
 
-    fun next(view: View) {
-        val continueButton: Button = findViewById(R.id.continue_button)
-        continueButton.setOnClickListener {
-            val intent = Intent(this, PetGenderActivity::class.java)
-            startActivity(intent)
+    fun confirm(view: View) {
+        val confirmButton: Button = findViewById(R.id.confirm_button)
+        confirmButton.setOnClickListener {
+            val isValid = ValidationObject.petNameValidation(findViewById(R.id.name_edit_text))
+            if (isValid) {
+                val intent = Intent(this, PetListFragment::class.java)
+                startActivity(intent)
+            }
         }
     }
