+++ app/src/main/java/com/petsup/ui/fragment/HomeFragment.kt
@@ -1,11 +1,15 @@
-package com.petsup.fragment
+package com.petsup.ui.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.petsup.R
+import androidx.recyclerview.widget.LinearLayoutManager
+import com.petsup.databinding.FragmentHomeBinding
+import com.petsup.models.Petshop
+import com.petsup.services.PetshopService
+import com.petsup.ui.adapter.PetshopsAdapter
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -17,10 +21,13 @@ private const val ARG_PARAM2 = "param2"
  * Use the [HomeFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class PetListFragment : Fragment() {
+class HomeFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
+
+    private lateinit var binding: FragmentHomeBinding
+    private lateinit var petshopService: PetshopService
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -34,8 +41,19 @@ class PetListFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_pet_list, container, false)
+        binding = FragmentHomeBinding.inflate(layoutInflater)
+        initRecyclerView()
+        return binding.root
+    }
+
+    private fun initRecyclerView() {
+        binding.recyclerView.layoutManager = LinearLayoutManager(context)
+        binding.recyclerView.setHasFixedSize(true)
+        binding.recyclerView.adapter = PetshopsAdapter(getPetshops())
+    }
+
+    private fun getPetshops(): List<Petshop> {
+        return petshopService.listPetshops().execute().body() ?: emptyList()
     }
 
     companion object {
