+++ app/src/main/res/layout/fragment_home.xml
@@ -56,99 +56,10 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/section_title">
 
-        <androidx.cardview.widget.CardView
-            android:id="@+id/pet_shop_card_view_1"
-            style="@style/default_card_view"
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/recycler_view"
             android:layout_width="match_parent"
-            android:layout_height="104dp"
-            android:layout_marginHorizontal="17dp"
-            android:layout_marginVertical="16dp"
-            android:minHeight="104dp">
-
-            <androidx.constraintlayout.widget.ConstraintLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent">
-
-                <ImageView
-                    android:id="@+id/pet_shop_icon_1"
-                    android:layout_width="80dp"
-                    android:layout_height="80dp"
-                    android:layout_marginStart="8dp"
-                    android:layout_marginTop="12dp"
-                    android:src="@drawable/petshop_icon"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="parent" />
-
-                <TextView
-                    android:id="@+id/card_title_1"
-                    style="@style/default_title_text_view"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="8dp"
-                    android:layout_marginTop="8dp"
-                    android:text="Fofinho Pet shop"
-                    app:layout_constraintStart_toEndOf="@id/pet_shop_icon_1"
-                    app:layout_constraintTop_toTopOf="parent" />
-
-                <TextView
-                    android:id="@+id/card_location_1"
-                    style="@style/gray_description_text_view"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="Cerqueira César, São Paulo"
-                    app:layout_constraintStart_toStartOf="@id/card_title_1"
-                    app:layout_constraintTop_toBottomOf="@id/card_title_1" />
-
-                <TextView
-                    android:id="@+id/card_services_1"
-                    style="@style/gray_description_text_view"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="@string/services"
-                    app:layout_constraintStart_toStartOf="@id/card_location_1"
-                    app:layout_constraintTop_toBottomOf="@id/card_location_1" />
-
-                <TextView
-                    android:id="@+id/card_status_1"
-                    style="@style/green_description_text_view"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="@string/status_open"
-                    app:layout_constraintStart_toStartOf="@id/card_services_1"
-                    app:layout_constraintTop_toBottomOf="@id/card_services_1" />
-
-                <ImageView
-                    android:id="@+id/grade_image_view_1"
-                    android:layout_width="16dp"
-                    android:layout_height="16dp"
-                    android:layout_marginTop="4dp"
-                    android:layout_marginEnd="2dp"
-                    android:src="@drawable/star_icon"
-                    app:layout_constraintEnd_toStartOf="@id/grade_text_view_1"
-                    app:layout_constraintTop_toTopOf="@id/card_title_1" />
-
-                <TextView
-                    android:id="@+id/grade_text_view_1"
-                    style="@style/grade_text_view"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="2dp"
-                    android:layout_marginEnd="8dp"
-                    android:text="4.5"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintTop_toTopOf="@id/card_title_1" />
-
-                <ImageView
-                    android:layout_width="24dp"
-                    android:layout_height="24dp"
-                    android:layout_marginTop="8dp"
-                    android:src="@drawable/favorite_icon"
-                    app:layout_constraintStart_toStartOf="@id/grade_text_view_1"
-                    app:layout_constraintTop_toBottomOf="@id/grade_text_view_1" />
-
-            </androidx.constraintlayout.widget.ConstraintLayout>
-
-        </androidx.cardview.widget.CardView>
+            android:layout_height="wrap_content" />
 
     </ScrollView>
